refactor(analytics): use shared reporting database
